Change the name of the int test project and align namespace changes
