+++ SourceFormats/SourceFormatsRepository.Int.Tests/SourceFormatNode/AddChildNodeValidationShould.cs
@@ -1,24 +1,28 @@
-namespace EncyclopediaGalactica.SourceFormats.Repository.Int.Tests.SourceFormatNode;
+namespace EncyclopediaGalactica.SourceFormats.SourceFormatsRepository.Int.Tests.SourceFormatNode;
 
 using System;
 using System.Diagnostics.CodeAnalysis;
 using System.Threading.Tasks;
+using EncyclopediaGalactica.SourceFormats.SourceFormatsRepository.Exceptions;
+using EncyclopediaGalactica.Utils.GuardsService;
 using FluentAssertions;
-using SourceFormatsRepository.Exceptions;
-using Utils.GuardsService;
 using Xunit;
 
 [ExcludeFromCodeCoverage]
 [Collection("SourceFormatNode Repository Collection")]
-public class GetByIdWithTreeValidationShould : BaseTest
+public class AddChildNodeValidationShould : BaseTest
 {
-    [Fact]
-    public async Task ThrowWhenInputIsInvalid()
+    [Theory]
+    [InlineData(0, 0)]
+    [InlineData(1, 0)]
+    [InlineData(0, 1)]
+    [InlineData(1, 1)]
+    public async Task ThrowWhenInputIsInvalid(long childId, long parentId)
     {
-        // Act
+        // act
         Func<Task> action = async () =>
         {
-            await Sut.SourceFormatNodes.GetByIdWithFlatTreeAsync(0).ConfigureAwait(false);
+            await Sut.SourceFormatNodes.AddChildNodeAsync(childId, parentId, parentId).ConfigureAwait(false);
         };
 
         // Assert
