+++ SourceFormats/SourceFormatsRepository.Int.Tests/SourceFormatNode/AddValidationShould.cs
@@ -1,12 +1,12 @@
-namespace EncyclopediaGalactica.SourceFormats.Repository.Int.Tests.SourceFormatNode;
+namespace EncyclopediaGalactica.SourceFormats.SourceFormatsRepository.Int.Tests.SourceFormatNode;
 
 using System;
 using System.Diagnostics.CodeAnalysis;
 using System.Threading.Tasks;
-using Entities;
+using EncyclopediaGalactica.SourceFormats.Entities;
+using EncyclopediaGalactica.SourceFormats.SourceFormatsRepository.Exceptions;
 using FluentAssertions;
 using FluentValidation;
-using SourceFormatsRepository.Exceptions;
 using Xunit;
 
 [ExcludeFromCodeCoverage]
