+++ SourceFormats/SourceFormatsRepository.Int.Tests/SourceFormatNode/GetByIdShould.cs
@@ -1,8 +1,8 @@
-namespace EncyclopediaGalactica.SourceFormats.Repository.Int.Tests.SourceFormatNode;
+namespace EncyclopediaGalactica.SourceFormats.SourceFormatsRepository.Int.Tests.SourceFormatNode;
 
 using System.Diagnostics.CodeAnalysis;
 using System.Threading.Tasks;
-using Entities;
+using EncyclopediaGalactica.SourceFormats.Entities;
 using FluentAssertions;
 using Xunit;
