+++ SourceFormats/SourceFormatsRepository.Int.Tests/SourceFormatNode/RepositoryCollectionDefinition.cs
@@ -1,4 +1,4 @@
-namespace EncyclopediaGalactica.SourceFormats.Repository.Int.Tests.SourceFormatNode;
+namespace EncyclopediaGalactica.SourceFormats.SourceFormatsRepository.Int.Tests.SourceFormatNode;
 
 using System.Diagnostics.CodeAnalysis;
 using Xunit;
